Enhanced CMake Support
The current repository exports it's targets using the usual regime, but
sets neither the include directories or alias targets. This means that,
while able to be consumed by projects using `find_package()`, it it
not able to be consumed properly by those using `ExternalProject`,
`FetchContent`, `CPM` or `add_subdirectory`. The include directories
allow such projects to properly locate the headers (however without
the usual `liquid/` prefix). The `ALIAS` targets simply allow the targets
to be linked against in a uniform way.

* Add include directories
* Add alias targets for liquid-shared and liquid-static